+++ src/console.c
@@ -1339,18 +1339,18 @@ static void render_sensor_object(struct console_request_context *ctx,
                          "Value to assume in fault condition",
                          &input->fault_config.fault_value);
 
-  render_uint32_map_field(ctx,
-                          "window-capture-width",
+  render_float_map_field(ctx,
+                          "window-capture-opening",
                           "Crank degrees in window to average samples over",
-                          &input->window.capture_width);
+                          &input->window.window_opening);
   render_uint32_map_field(ctx,
-                          "window-total-width",
-                          "Crank degrees per window",
-                          &input->window.total_width);
-  render_uint32_map_field(ctx,
+                          "window-count",
+                          "windows per engine cycle",
+                          &input->window.windows_per_cycle);
+  render_float_map_field(ctx,
                           "window-offset",
                           "Crank degree into window to start averagine",
-                          &input->window.offset);
+                          &input->window.window_offset);
 }
 
 static void render_knock_object(struct console_request_context *ctx,